Added mutex and event scheduling function to midi_hw_if
Almost done the CoreMIDI implementation, there are just a few unknowns.  We want
the caller to have an abstracted interface, that is, they should have to call
the driver for their system (OSX calls CoreMIDI, others Jack, etc.) but after
that they should have an abstracted interface. Here's what I think they might
want in an interface:

A way to set a function that will be passed midi_events in order to inform the
program of what events were received. No midi receive functionality has yet been
added to the midi_hw_if, but when it is, this is probably how it will be
accessed.

A way to pass in a function that can schedule MIDI events to play (this already
exists).

A way to get the current time. This is currently not part of the midi_hw_if and
so would have to be wrapped up in a struct of some kind if you want to keep
these together.

A possibility would be that the hardware returns function
pointers that the caller can use to get the time, schedule MIDI, get MIDI
events, etc. These pointers could even get returned to a struct to keep them
together.